% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_esr_template.R
\name{build_esr_template}
\alias{build_esr_template}
\title{Build Ecosystem Summary Report Template}
\usage{
build_esr_template(
  region,
  ps_vars = c("nao", "amo", "azmp_satellite_temperature", "azmp_surface_temperature",
    "azmp_bottom_temperature", "azmp_stratification", "azmp_salinity",
    "ocean_acidification", "future_projections"),
  ecol_vars = c("commercial_fishing", "fishing_pressure", "cumulative_impacts",
    "nitrate", "chlorophyll", "bloom_stat", ".c_finmarchicus_template", "zoo_biomass",
    "large_fish", "biodiversity_vars", "guild_biomass", "nontarget_species"),
  econ_vars = c("commercial_fisheries_value"),
  out.dir = "template_reports"
)
}
\arguments{
\item{region}{the region for which to build template report. Options are: "4X", "4W", "4V", "4VN", "4VS", "ESS", "WSS", or "blank" for a blank template report}

\item{ps_vars}{Pressures and Stressors variables to include in a report. Options are: "nao", "amo", "azmp_satellite_temperature", "azmp_surface_temperature", "azmp_bottom_temperature", "azmp_stratification", "azmp_salinity", "ocean_acidification", "future_projections". Defaults to all.}

\item{ecol_vars}{Ecological Variables to include in report template. Options are "commercial_fishing", "fishing_pressure", "cumulative_impacts", "nitrate", "chlorophyll", "bloom_stat", ".c_finmarchicus_template", "zoo_biomass", "large_fish", "biodiversity_vars", "guild_biomass", "nontarget_species". Defaults to all.}

\item{econ_vars}{Economic Variabiles to include in report template. Options are "commercial_fisheries_value". Defaults to all.}

\item{out.dir}{the directory in which to save a generated report.}
}
\value{
a pre-filled .Rmd document containing script to generate a template Ecosystem Summary Report
}
\description{
Build Ecosystem Summary Report Template
}
\examples{
build_esr_template(region = "4W")
}
