% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.1_rmd_header.R
\name{rmd_header}
\alias{rmd_header}
\title{Create Report Header}
\usage{
rmd_header(.region)
}
\arguments{
\item{.region}{Region for report; inherited from `build_esr_template()`.}
}
\value{
Header section for report: Rmd header, and code chunk for required packages.
}
\description{
Create Report Header
}
